refactor : adding Exception account.isPresent

+++ src/main/java/com/daelim/yourmind/emotion/service/EmotionServiceImpl.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Objects;
+import java.util.Optional;
 import java.util.function.Function;
 
 
@@ -27,73 +28,96 @@ public class EmotionServiceImpl implements EmotionService {
     @Override
     public StatusDTO saveEmotion(EmotionDTO emotionDTO) {
         try {
-            User child = userRepository.findByUsername(emotionDTO.getChild());
-            User counselor = userRepository.findByUsername(emotionDTO.getCounselor());
-            Emotion emotion = dtoToEntity(emotionDTO, child, counselor);
-            emotionRepository.save(emotion).getId();
-            StatusDTO statusDTO = StatusDTO.builder().status("success").build();
-            return statusDTO;
+            Optional<User> childOption = userRepository.findByUsername(emotionDTO.getChild());
+            Optional<User> counselorOption = userRepository.findByUsername(emotionDTO.getCounselor());
+            if (childOption.isPresent() && counselorOption.isPresent()) {
+                User child = childOption.get();
+                User counselor = counselorOption.get();
+                Emotion emotion = dtoToEntity(emotionDTO, child, counselor);
+                emotionRepository.save(emotion);
+                StatusDTO statusDTO = StatusDTO.builder().status("success").build();
+                return statusDTO;
+            } else {
+                throw new RuntimeException("This account doesn't exist");
+            }
         } catch(Exception e) {
+            e.printStackTrace();
             throw e;
         }
     }
 
     @Override
     public EmotionDTO getEmotion(Long emotionId, String username) {
-        Emotion emotion = emotionRepository.findById(emotionId).get();
-        if (emotion.getChild().getUsername().equals(username) || emotion.getCounselor().getUsername().equals(username)) {
-            User child = emotion.getChild();
-            User counselor = emotion.getCounselor();
-            return entityToDTO(emotion, child, counselor);
+        Optional<Emotion> emotionOption = emotionRepository.findById(emotionId);
+        Emotion emotion;
+        if (emotionOption.isPresent()) {
+            emotion = emotionOption.get();
         } else {
-            throw new RuntimeException("No permission");
+            throw new RuntimeException("Not found");
+        }
+        Optional<User> userOptional = userRepository.findByUsername(username);
+        if (userOptional.isPresent()) {
+            if (emotion.getChild().getUsername().equals(username) || emotion.getCounselor().getUsername().equals(username)) {
+                User child = emotion.getChild();
+                User counselor = emotion.getCounselor();
+                return entityToDTO(emotion, child, counselor);
+            } else {
+                throw new RuntimeException("No permission");
+            }
+        } else {
+            throw new RuntimeException("This account doesn't exist");
         }
     }
 
     @Override
-    @Transactional
     public PageResultDTO<EmotionDTO, Object[]> getEmotions(PageRequestDTO pageRequestDTO) {
         log.info(pageRequestDTO);
         Function<Object[], EmotionDTO> fn = (
-                entity -> entityToDTO((Emotion)entity[0],
+                entity -> entityToDTO(
+                        (Emotion)entity[0],
                         (User)entity[1],
                         (User)entity[2])
         );
         Page<Object[]> result;
         String username = pageRequestDTO.getUsername();
-        if (userRepository.existsByUsername(username)) {
+        if (userRepository.findByUsername(username).isPresent()) {
             if (isCounselor(username)) {
                 result = emotionRepository.getEmotionAndChildAndCounselorByCounselor(
                     pageRequestDTO.getPageable(Sort.by("id").descending()),
-                    userRepository.findByUsername(username).getId()
+                    userRepository.findByUsername(username).get().getId()
                 );
             } else {
                 result = emotionRepository.getEmotionAndChildAndCounselorByChild(
                     pageRequestDTO.getPageable(Sort.by("id").descending()),
-                    userRepository.findByUsername(username).getId()
+                    userRepository.findByUsername(username).get().getId()
                 );
             }
             return new PageResultDTO<>(result, fn);
         } else {
-            throw new RuntimeException("No permission");
+            throw new RuntimeException("This account doesn't exist");
         }
     }
 
     @Override
     public StatusDTO deleteEmotion(Long emotionId, String username) {
-        Emotion emotion = emotionRepository.findById(emotionId).get();
-        if (emotion.getChild().getUsername().equals(username) || emotion.getCounselor().getUsername().equals(username)) {
-            emotionRepository.deleteById(emotionId);
-            StatusDTO dto = StatusDTO.builder().status("success").build();
-            return dto;
+        Optional<Emotion> emotionOption = emotionRepository.findById(emotionId);
+        if (emotionOption.isPresent()) {
+            Emotion emotion = emotionRepository.findById(emotionId).get();
+            if (emotion.getChild().getUsername().equals(username) || emotion.getCounselor().getUsername().equals(username)) {
+                emotionRepository.deleteById(emotionId);
+                StatusDTO dto = StatusDTO.builder().status("success").build();
+                return dto;
+            } else {
+                throw new RuntimeException("No permission");
+            }
         } else {
-            throw new RuntimeException("No permission");
+            throw new RuntimeException("This account doesn't exist");
         }
     }
 
     public boolean isCounselor(String username) {
-        User user = userRepository.findByUsername(username);
-        return user.isCounselor();
+        Optional<User> user = userRepository.findByUsername(username);
+        return user.get().isCounselor();
     };
 
 }
